requirements.txt
boto3==1.28.17
botocore==1.31.17
fake-useragent==1.1.3
pandas==2.0.3
psycopg2-binary==2.9.7
pytz==2022.6
requests==2.31.0
pyarrow==12.0.1